consolidate ClientAuthenticationWithToken and fix tests
