+++ e2e/test/iothub/DeviceTokenRefreshE2ETests.cs
@@ -112,7 +112,7 @@ namespace Microsoft.Azure.Devices.E2ETests
                 Target = $"{iotHub}/devices/{WebUtility.UrlEncode(deviceId)}",
             };
 
-            var auth = new DeviceAuthenticationWithToken(deviceId, builder.ToSignature());
+            var auth = new ClientAuthenticationWithToken(builder.ToSignature(), deviceId);
 
             using var deviceClient = new IotHubDeviceClient(iotHub, auth, new IotHubClientOptions(new IotHubClientAmqpSettings()));
             Logger.Trace($"{deviceId}: Created {nameof(IotHubDeviceClient)} ID={TestLogger.IdOf(deviceClient)}");
